refactor: host sidebar item UI inside the item's list element
Developer entries rendered their dialog as a sibling of NavItem, which
put it directly inside the menu's ul. That only worked because dialogs
portal their content out, and nothing stopped a future entry from
emitting inline markup there.

Let NavItem take children and host them in the SidebarMenuItem it
already renders, so the item's own li is the anchor and the constraint
is structural rather than a comment.

Also move sidebarDeveloperGroup next to its entry hooks. Naming a hook
from config/ dragged the store and the editor loader into a module whose
other consumers only want icons and routes; the config stays
declarative now.
